refactor GFlowsLib
