+++ workflow/engine/jsonnet_template_engine.go
@@ -12,9 +12,9 @@ import (
 
 	gojsonnet "github.com/google/go-jsonnet"
 	"github.com/jbrunton/gflows/config"
+	"github.com/jbrunton/gflows/env"
 	"github.com/jbrunton/gflows/io"
 	"github.com/jbrunton/gflows/workflow/engine/jsonnet"
-	"github.com/jbrunton/gflows/workflow/env"
 	"github.com/spf13/afero"
 )
